refactor functionality to helper classes
OCD-2159
