+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/changerequest/manager/ChangeRequestManager.java
@@ -4,7 +4,7 @@ import gov.healthit.chpl.changerequest.domain.ChangeRequest;
 import gov.healthit.chpl.exception.EntityRetrievalException;
 
 public interface ChangeRequestManager {
-    ChangeRequest createWebsiteChangeRequest(final ChangeRequest cr)
+    ChangeRequest createChangeRequest(final ChangeRequest cr)
             throws EntityRetrievalException;
 
     ChangeRequest getChangeRequest(final Long changeRequestId) throws EntityRetrievalException;
